Resolve circular dep with StrideRange (#1818)
