+++ pmlc/dialect/pxa/analysis/strides.cc
@@ -1,4 +1,4 @@
-// Copyright 2020 Intel Corporation
+// Copyright 2020, Intel Corporation
 
 #include "pmlc/dialect/pxa/analysis/strides.h"
 
@@ -75,60 +75,6 @@ static Optional<StrideInfo> flatten(MemRefType memRefType,
   return flat;
 }
 
-StrideRange::StrideRange(BlockArgument arg)
-    : valid(false), minVal(0), maxVal(0), stride(0) {
-  if (auto ap = dyn_cast<AffineParallelOp>(arg.getOwner()->getParentOp())) {
-    auto rangeExpr = util::getRangesValueMap(ap).getResult(arg.getArgNumber());
-    auto rangeConstantExpr = rangeExpr.dyn_cast<AffineConstantExpr>();
-    if (!rangeConstantExpr) {
-      return;
-    }
-    int64_t range = rangeConstantExpr.getValue();
-    if (range < 1) {
-      return;
-    }
-    auto steps = ap.getSteps();
-    int64_t step = steps[arg.getArgNumber()];
-    if (step <= 0) {
-      return;
-    }
-    stride = 1;
-    minVal = 0;
-    // This is a correction to deal with the fact that strides are measured
-    // relative to loop iterations not indexes.
-    maxVal = (range - 1) / step;
-    valid = true;
-    if (minVal == maxVal) {
-      stride = 0;
-    }
-  }
-}
-
-StrideRange &StrideRange::operator*=(int64_t factor) {
-  minVal *= factor;
-  maxVal *= factor;
-  stride *= factor;
-  if (factor < 0) {
-    std::swap(minVal, maxVal);
-  }
-  return *this;
-}
-
-StrideRange &StrideRange::operator+=(const StrideRange &rhs) {
-  valid = valid && rhs.valid;
-  minVal += rhs.minVal;
-  maxVal += rhs.maxVal;
-  stride = std::gcd(stride, rhs.stride);
-  return *this;
-}
-
-void StrideRange::unionEquals(const StrideRange &rhs) {
-  valid = valid && rhs.valid;
-  minVal = std::min(minVal, rhs.minVal);
-  maxVal = std::max(maxVal, rhs.maxVal);
-  stride = std::gcd(stride, rhs.stride);
-}
-
 // Multiply the offset and all strides by a constant.
 StrideInfo &StrideInfo::operator*=(int64_t factor) {
   offset *= factor;
@@ -156,11 +102,6 @@ StrideInfo &StrideInfo::operator+=(const StrideInfo &rhs) {
     }
   }
   return *this;
-}
-
-std::ostream &operator<<(std::ostream &os, const StrideRange &val) {
-  os << '(' << val.minVal << ", " << val.maxVal << "]:" << val.stride;
-  return os;
 }
 
 static BoundaryRegion getBoundaryRegion(Block *x, Block *y) {
