+++ pmlc/dialect/pxa/analysis/strides.h
@@ -8,8 +8,8 @@
 #include "llvm/ADT/Optional.h"
 
 #include "pmlc/dialect/pxa/analysis/affine_expr.h"
-#include "pmlc/dialect/pxa/analysis/stride_range.h"
 #include "pmlc/dialect/pxa/ir/ops.h"
+#include "pmlc/dialect/pxa/ir/stride_range.h"
 
 namespace pmlc::dialect::pxa {
